pyproject.toml
[project]
name = "srspipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "fastapi>=0.126.0",
    "google-genai>=1.56.0",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.38.0",
]
